# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book Store API",
        "description": "Book Store API",
        "version": "1.0.0",
        "contact": {
            "name": "Bikash Chauhan [20BCE2769]",
            "url": "https://github.com/bikash-789/BookStore_API"
        }
    },
    "host": "localhost:3000",
    "basePath": "/",
    "tags": [
    {
      "name": "book-store",
      "description": "Endpoints related to the book store"
    },
    {
        "name": "user",
        "description": "Endpoints related to the user"
    }
  ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": ["user"],
                "summary": "User login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": { "type": "string" },
                                    "password": { "type": "string" }
                                },
                                "required": ["email", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "access_token": { "type": "string" }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Invalid credentials."
                    }
                }
            }
        },
        "/api/signup": {
            "post": {
                "tags": ["user"],
                "summary": "User signup",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": { "type": "string" },
                                    "password": { "type": "string" }
                                },
                                "required": ["email", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Signup successful"
                    },
                    "400": {
                        "description": "Bad Request. Invalid input."
                    }
                }
            }
        },
        "/api/books": {
            "get": {
                "tags": ["book-store"],
                "summary": "Get a list of books",
                "responses": {
                    "200": {
                        "description": "A list of books in the database",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "author": {
                                                "type": "string"
                                            },
                                            "ISBN": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "integer"
                                            },
                                            "quantity": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Books not found!"
                    },
                    "500": {
                        "description": "Application failed to process the request"
                    }
                }
            },
            "post": {
                "tags": ["book-store"],
                "summary": "Add new books",
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "ISBN": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    }
                                },
                                "required": [
                                    "title",
                                    "author",
                                    "ISBN",
                                    "price",
                                    "quantity"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Books added successfully"
                    },
                    "401": {
                        "description": "Unauthorized. Authentication required."
                    },
                    "500": {
                        "description": "Application failed to process the request"
                    }
                }
            }
        },
        "/api/book": {
            "post": {
                "tags": ["book-store"],
                "summary": "Add a new book",
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "ISBN": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    }
                                },
                                "required": [
                                    "title",
                                    "author",
                                    "ISBN",
                                    "price",
                                    "quantity"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book added successfully"
                    },
                    "401": {
                        "description": "Unauthorized. Authentication required."
                    },
                    "500": {
                        "description": "Application failed to process the request"
                    }
                }
            }
        },
        "/api/book/{isbn}": {
            "get": {
                "tags": ["book-store"],
                "summary": "Get a book by ISBN",
                "parameters": [
                    {
                        "name": "isbn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "title": {
                                            "type": "string"
                                        },
                                        "author": {
                                            "type": "string"
                                        },
                                        "ISBN": {
                                            "type": "string"
                                        },
                                        "price": {
                                            "type": "integer"
                                        },
                                        "quantity": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found"
                    },
                    "500": {
                        "description": "Application failed to process the request"
                    }
                }
            },
            "put": {
                "tags": ["book-store"],
                "summary": "Update a book by ISBN",
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "name": "isbn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized. Authentication required."
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "delete": {
                "tags": ["book-store"],
                "summary": "Delete a book by ISBN",
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "name": "isbn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized. Authentication required."
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}